Write array_merge function (details in description)
The array_merge function calls the is_unique function to create the binaryArray, in which numbers at indexes corresponding to values are set to 1 if these values occur in the input array of arrays. The binaryArray's last index is reserved for a tally of the total unique occurrences of values. By accessing this tally, we are able to create a new array, mergedArray, for our results. Thus, using the information in the binary array, we are assign values to the result array.